requirements.txt
openai==2.15.0
numpy==2.0.2
lightning==2.6.0
pytorch-lightning==2.6.0
transformers==4.57.6
soundfile==0.13.1
einops==0.8.1
librosa==0.11.0
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
pydub==0.25.1
PyYAML==6.0.2
# replicate==0.33.0
requests==2.32.3
# resampy==0.4.3
spacy==3.8.11
streamlit==1.53.0
yt-dlp
json-repair
ruamel.yaml
InquirerPy
autocorrect-py
ctranslate2==4.6.3
edge-tts
streamlit-authenticator==0.4.2
whisperx @ git+https://github.com/m-bain/whisperx.git@d32ec3e3012ec4c0934f4088424c32f3f038b249
rich
syllables
pypinyin
g2p-en
pydantic==2.12.5

